requirements.txt
annotated-types==0.6.0
asciitree==0.3.3
asttokens==2.4.1
attrs==23.2.0
beautifulsoup4==4.12.3
certifi==2023.11.17
charset-normalizer==3.3.2
clarabel==0.6.0
click==8.1.7
cloudpickle==3.0.0
colorama==0.4.6
comm==0.2.1
contourpy==1.2.0
cramjam==2.8.0
cvxpy==1.4.1
cycler==0.12.1
dask==2024.1.0
debugpy==1.8.0
decorator==5.1.1
dice-ml==0.11
distributed==2024.1.0
ecos==2.0.12
executing==2.0.1
fasteners==0.19
fastparquet==2023.10.1
filelock==3.13.1
fonttools==4.47.2
fsspec==2023.12.2
google==3.0.0
grpcio==1.60.1
grpcio-tools==1.60.1
htmlmin==0.1.12
idna==3.6
ImageHash==4.3.1
importlib-metadata==7.0.1
ipykernel==6.29.0
ipython==8.20.0
jedi==0.19.1
Jinja2==3.1.3
joblib==1.1.1
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter_client==8.6.0
jupyter_core==5.7.1
kiwisolver==1.4.5
locket==1.0.0
MarkupSafe==2.1.3
matplotlib==3.8.2
matplotlib-inline==0.1.6
maxminddb==2.5.2
maxminddb-geolite2==2018.703
missingno==0.5.2
mpmath==1.3.0
msgpack==1.0.7
multimethod==1.11
nest-asyncio==1.5.9
networkx==3.2.1
numcodecs==0.12.1
numpy==1.26.3
osqp==0.6.3
packaging==23.2
pandas==1.5.3
pandas-profiling==3.2.0
parso==0.8.3
partd==1.4.1
patsy==0.5.6
phik==0.12.4
pillow==10.2.0
platformdirs==4.1.0
prompt-toolkit==3.0.43
protobuf==4.25.2
psutil==5.9.7
pure-eval==0.2.2
PyALE==1.1.3
pyaml==23.12.0
pyarrow==14.0.2
pybind11==2.11.1
pydantic==2.6.1
pydantic_core==2.16.2
pyDeprecate==0.3.2
pyDVL==0.8.0
Pygments==2.17.2
pymemcache==4.0.0
pyparsing==3.1.1
python-dateutil==2.8.2
pytz==2023.3.post1
PyWavelets==1.5.0
pywin32==306
PyYAML==6.0.1
pyzmq==25.1.2
qdldl==0.1.7.post0
raiutils==0.4.1
referencing==0.33.0
requests==2.31.0
rpds-py==0.17.1
scikit-learn==1.3.2
scikit-optimize==0.9.0
scipy==1.11.4
scs==3.2.4.post1
seaborn==0.13.1
six==1.16.0
skorch==0.15.0
sortedcontainers==2.4.0
soupsieve==2.5
stack-data==0.6.3
statsmodels==0.14.1
sympy==1.12
tabulate==0.9.0
tangled-up-in-unicode==0.2.0
tblib==3.0.0
threadpoolctl==3.2.0
toolz==0.12.0
torch==2.1.2
torchvision==0.16.2
tornado==6.4
tqdm==4.66.1
traitlets==5.14.1
typing_extensions==4.9.0
tzdata==2023.4
urllib3==2.1.0
visions==0.7.4
wcwidth==0.2.13
xgboost==2.0.3
zarr==2.16.1
zict==3.0.0
zipp==3.17.0
